remove `export` form the beginning of `.env` files to make it compatible with shell environments
fixes #102

A simpler regex would be `"^\\s*export\\s+(?=\\w)"` but there is a pathological case, if an environment variable  is called `export` and there is a space before the value ('export = foobar`).  That's why I require the next character after the whitespace to be a `word` character (using he lookahead operator `(?=X)`.